[IMP] hr_holidays: add holidays data mobile overview
Issues:
- Mobile summary view now consistently lists all available time off types that have an allocation (>0 days). Additionally, it cleans up the mobile ui by removing redundant views of balance information into a single section.
- Minor problem in visualization of Summary mobile dropdown menu, folding on mobile screens when the bar was filled with entries along the whole width of the view

Removed # of available days from:
- the filter panel section (calendar_filter_section)
- the upper calendar bar (calendar_mobile_filter_panel)

Added:
- a Time Off Balance section in the calendar_side_panel

Task ID: 5030732
